enable use of iidfile for remote build
podman/buildah build allows for --iidfile which stores the id of the newly built image.  the api more or less was able to handle this but it wasnt being sent by the go bindings.  it also was not documented on the libpod side of the documentation because of a copy and paste with compat.

Signed-off-by: Brent Baude <bbaude@redhat.com>