HACK: Manually include c/storage #698
We need this release out by end of day, so we don't have time to
do this right. Disable the vendor task and manually add c/storage
PR #698 to the vendored copy of c/storage to make the tests pass.

Once #698 merges into c/storage, we need to remove this commit
and backport it to the v1.20 stable branch, then cut a release
there.

Signed-off-by: Matthew Heon <mheon@redhat.com>