Dump default AppArmor profile
Extend the `info` command to dump the default AppArmor profile used by
podman.

This can be useful to debug AppArmor related issues when the default
profile is too tight for the containerized application.

The default AppArmor profile is dynamically generated by podman at
runtime and, once loaded into the kernel, there's no way to print it
in a human-readable form.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>