Makefile: Fix broken libpodimage targets
The configuration for this container has moved to the
`containers/automation_images` repository.  Leverage the
image built from the new location to guarantee synchronization
with images used in CI.

Also, remove several other targets that haven't worked for a really
long time.  Upon discussion, they seem to apply to a minority of
users.  Remove them to clean up the `Makefile` and reduce
maintenance burden.

Signed-off-by: Chris Evich <cevich@redhat.com>