[PATCH] store: keep graph lock during Mount
This solves a race condition where a mountpoint is created without the
home mount being present.

The cause is that another process could be calling the graph driver
cleanup as part of store.Shutdown() causing the unmount of the
driver home directory.

The unmount could happen between the time the rlstore is retrieved and
the actual mount, causing the driver mount to be done without a home
mount below it.

A third process then would re-create again the home mount, shadowing
the previous mount.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1757845

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

<MH: Tweaked to apply against 1.6.4 branch>

Signed-off-by: Matthew Heon <mheon@redhat.com>