pkg/registries: add a retiring note
The registries package should be retired.  It was introduced as an
easier to use wrapper around c/image `sysregistries` which has been
replaced by `sysregistriesv2` a long while ago.

Users should either use the `sysregistriesv2` package directly or, even
better, we cache the config in libpod's image runtime to prevent
redundant (and ~expensive) parsing of the registries.conf files.

For now, just add a note in hope we'll not forgert about it when we find
time in the future.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>